# backend/docs/swagger.yaml
{
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Authentication, product, inventory, sales, branch, employee, dashboard, and template-style hexagonal endpoints.",
    "title": "Multi-Branch Inventory \u0026 POS API",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/swagger/doc.json": {
      "get": {
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "summary": "OpenAPI JSON document",
        "tags": [
          "Dashboard"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "Local Docker / Nginx",
      "url": "http://localhost:8080"
    },
    {
      "description": "Local frontend proxy",
      "url": "http://localhost:3000"
    }
  ],
...
}